requirements.txt
python-dotenv
streamlit
snowflake-snowpark-python[pandas]
py3Dmol
stmol
pandas
ipython_genutils